Fix bugs from refactoring Epifany to algorithm class. Revert change preventing compilation on gcc 4.8. Add test for the Epifany tool. Change protein scores to doubles for extended precision in FDR ranking.

+++ src/openms/source/METADATA/ProteinHit.cpp
@@ -90,7 +90,7 @@ namespace OpenMS
 
   //TODO WHAAAAAT why float???
   // returns the score of the protein hit
-  float ProteinHit::getScore() const
+  double ProteinHit::getScore() const
   {
     return score_;
   }
